+++ packages/desktop/src/js/hash.txt
@@ -1,0 +1,1 @@
+[11927251734412729446]
