+++ packages/desktop/src/js/native_eval.js
