+++ packages/html/src/lib.rs
@@ -27,14 +27,8 @@ pub use file_data::*;
 mod attribute_groups;
 pub mod geometry;
 pub mod input_data;
-#[cfg(feature = "native-bind")]
-pub mod native_bind;
 pub mod point_interaction;
 mod render_template;
-#[cfg(feature = "wasm-bind")]
-mod web_sys_bind;
-#[cfg(feature = "wasm-bind")]
-pub use web_sys_bind::*;
 
 #[cfg(feature = "serialize")]
 mod transit;
