--- packages/html/src/native_bind/mod.rs
@@ -1,3 +1,0 @@
-mod native_file_engine;
-
-pub use native_file_engine::*;
