--- packages/html/src/web_sys_bind/mod.rs
@@ -1,5 +1,0 @@
-mod events;
-#[cfg(feature = "file_engine")]
-mod file_engine;
-#[cfg(feature = "file_engine")]
-pub use file_engine::*;
