+++ packages/web/Cargo.toml
@@ -12,7 +12,7 @@ keywords = ["dom", "ui", "gui", "react", "wasm"]
 [dependencies]
 dioxus-core = { workspace = true }
 dioxus-core-types = { workspace = true }
-dioxus-html = { workspace = true, features = ["wasm-bind"] }
+dioxus-html = { workspace = true }
 dioxus-devtools = { workspace = true }
 dioxus-signals = { workspace = true }
 dioxus-interpreter-js = { workspace = true, features = [
@@ -38,33 +38,66 @@ serde = { version = "1.0", optional = true }
 serde-wasm-bindgen = { version = "0.5.0", optional = true }
 
 ciborium = { workspace = true, optional = true }
+async-trait = { version = "0.1.58", optional = true }
 
 [dependencies.web-sys]
 version = "0.3.56"
 features = [
+    "AnimationEvent",
+    "ClipboardEvent",
+    "CloseEvent",
+    "Comment",
+    "CompositionEvent",
+    "console",
+    "CustomEvent",
+    "DataTransfer",
     "Document",
+    "DragEvent",
+    "FocusEvent",
     "HtmlElement",
+    "HtmlFormElement",
     "HtmlInputElement",
     "HtmlSelectElement",
     "HtmlTextAreaElement",
-    "HtmlFormElement",
-    "Text",
-    "Comment",
-    "Window",
-    "DataTransfer",
-    "console",
+    "InputEvent",
+    "KeyboardEvent",
+    "MouseEvent",
     "NodeList",
-    "CloseEvent",
-    "CustomEvent",
+    "PointerEvent",
+    "ResizeObserverEntry",
+    "ResizeObserverSize",
+    "Text",
+    "Touch",
+    "TouchEvent",
+    "TouchList",
+    "TransitionEvent",
+    "WheelEvent",
+    "Window",
 ]
+
+[build-dependencies]
+lazy-js-bundle = { workspace = true }
 
 [features]
 default = ["panic_hook", "mounted", "file_engine", "devtools", "document"]
 panic_hook = ["dep:console_error_panic_hook"]
 hydrate = ["web-sys/Comment", "ciborium", "dep:serde"]
-mounted = ["web-sys/Element", "dioxus-html/mounted"]
+mounted = [
+    "web-sys/Element",
+    "dioxus-html/mounted",
+    "web-sys/Element",
+    "web-sys/DomRect",
+    "web-sys/ScrollIntoViewOptions",
+    "web-sys/ScrollLogicalPosition",
+    "web-sys/ScrollBehavior",
+    "web-sys/HtmlElement",
+]
 file_engine = [
     "dioxus-html/file_engine",
+    "dep:async-trait",
+    "web-sys/File",
+    "web-sys/FileList",
+    "web-sys/FileReader"
 ]
 devtools = ["web-sys/MessageEvent", "web-sys/WebSocket", "web-sys/Location", "dep:serde_json", "dep:serde", "dioxus-core/serialize"]
 document = ["dioxus-html/document", "dep:serde-wasm-bindgen", "dep:serde_json", "dep:serde"]
