+++ packages/web/src/file_engine.rs
@@ -1,6 +1,6 @@
 use std::any::Any;
 
-use crate::FileEngine;
+use dioxus_html::FileEngine;
 use futures_channel::oneshot;
 use js_sys::Uint8Array;
 use wasm_bindgen::{prelude::Closure, JsCast};
